refactor(web): remove separate sync fs polyfill and use unified entry
