+++ packages/utoo-web/src/webpackLoaders/polyfills/nodePolyFills.ts
@@ -1,4 +1,4 @@
-import * as fs from "./fsPolyfill/sync";
+import * as fs from "./fsPolyfill";
 import * as workerThreads from "./workerThreadsPolyfill";
 
 const buffer = require("buffer");
